Allow 'summary' and 'description' in reference nodes
Allow 'summary' and 'description' properties in references nodes ($ref).
This should fix some conflicts with the spec (see GH#7).

Spec: http://spec.openapis.org/oas/v3.1.0#reference-object

As a non-standard extension we also allow '$comment', just in case that
is useful to users. Anyway, that's not our job to fully validate a spec,
so any invalid use of $comment should be caught by that other tool.